Applying "Effective Dart: Style" to files
